requirements.txt
Flask==2.3.3
python-dotenv
supabase
gotrue
python-dotenv==1.0.0
supabase==1.0.4
google-generativeai==0.3.2
Werkzeug==2.3.7
gunicorn==21.2.0
requests==2.31.0
